+++ src/main.rs
@@ -1,0 +1,9 @@
+mod arithmetic;
+mod elgamal;
+mod primitives;
+mod zkps;
+mod proved;
+mod simple;
+mod distributed;
+#[cfg(test)]
+mod tests;
